+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* Fix "view previous" not working as expected when shuffle is enabled
+
 ## [v1.41.0] - 2026-04-18
 [v1.41.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.41.0